+++ pytest_playwright/sync.py
@@ -22,12 +22,7 @@ from typing import (
     Dict,
     Generator,
     List,
-    Literal,
     Optional,
-    Protocol,
-    Sequence,
-    Union,
-    Pattern,
     cast,
 )
 
@@ -40,11 +35,6 @@ from playwright.sync_api import (
     Page,
     Playwright,
     sync_playwright,
-    ProxySettings,
-    StorageState,
-    HttpCredentials,
-    Geolocation,
-    ViewportSize,
 )
 from slugify import slugify
 import tempfile
@@ -53,49 +43,8 @@ from pytest_playwright.pytest_playwright import (
     _create_guid,
     _is_debugger_attached,
     _truncate_file_name,
+    CreateContextCallback,
 )
-
-
-class CreateContextCallback(Protocol):
-    def __call__(
-        self,
-        viewport: Optional[ViewportSize] = None,
-        screen: Optional[ViewportSize] = None,
-        no_viewport: Optional[bool] = None,
-        ignore_https_errors: Optional[bool] = None,
-        java_script_enabled: Optional[bool] = None,
-        bypass_csp: Optional[bool] = None,
-        user_agent: Optional[str] = None,
-        locale: Optional[str] = None,
-        timezone_id: Optional[str] = None,
-        geolocation: Optional[Geolocation] = None,
-        permissions: Optional[Sequence[str]] = None,
-        extra_http_headers: Optional[Dict[str, str]] = None,
-        offline: Optional[bool] = None,
-        http_credentials: Optional[HttpCredentials] = None,
-        device_scale_factor: Optional[float] = None,
-        is_mobile: Optional[bool] = None,
-        has_touch: Optional[bool] = None,
-        color_scheme: Optional[
-            Literal["dark", "light", "no-preference", "null"]
-        ] = None,
-        reduced_motion: Optional[Literal["no-preference", "null", "reduce"]] = None,
-        forced_colors: Optional[Literal["active", "none", "null"]] = None,
-        accept_downloads: Optional[bool] = None,
-        default_browser_type: Optional[str] = None,
-        proxy: Optional[ProxySettings] = None,
-        record_har_path: Optional[Union[str, Path]] = None,
-        record_har_omit_content: Optional[bool] = None,
-        record_video_dir: Optional[Union[str, Path]] = None,
-        record_video_size: Optional[ViewportSize] = None,
-        storage_state: Optional[Union[StorageState, str, Path]] = None,
-        base_url: Optional[str] = None,
-        strict_selectors: Optional[bool] = None,
-        service_workers: Optional[Literal["allow", "block"]] = None,
-        record_har_url_filter: Optional[Union[str, Pattern[str]]] = None,
-        record_har_mode: Optional[Literal["full", "minimal"]] = None,
-        record_har_content: Optional[Literal["attach", "embed", "omit"]] = None,
-    ) -> BrowserContext: ...
 
 
 class PytestPlaywrightSync(PytestPlaywright):
