+++ tests/test_sync.py
@@ -804,7 +804,7 @@ def test_artifact_collection_should_work_for_manually_created_contexts_keep_open
     testdir.makepyfile(
         """
         import pytest
-        from pytest_playwright.sync import CreateContextCallback
+        from pytest_playwright.pytest_playwright import CreateContextCallback
 
         def test_artifact_collection(browser, page, new_context: CreateContextCallback):
             page.goto("data:text/html,<div>hello</div>")
